+++ src/main/kotlin/org/bibletranslationtools/fetcher/web/controllers/HomeController.kt
@@ -1,0 +1,38 @@
+package org.bibletranslationtools.fetcher.web.controllers
+
+import dev.jbs.ktor.thymeleaf.ThymeleafContent
+import io.ktor.application.ApplicationCallPipeline
+import io.ktor.application.call
+import io.ktor.request.acceptLanguage
+import io.ktor.request.uri
+import io.ktor.response.respond
+import io.ktor.routing.Routing
+import io.ktor.routing.get
+import io.ktor.routing.route
+import java.util.Locale
+import org.bibletranslationtools.fetcher.usecase.DependencyResolver
+import org.bibletranslationtools.fetcher.web.controllers.utils.getPreferredLocale
+
+fun Routing.homeController(resolver: DependencyResolver) {
+    route("/") {
+        var contentLanguage = listOf<Locale.LanguageRange>()
+        // execute before any sub-routes
+        intercept(ApplicationCallPipeline.Call) {
+            if (!call.request.uri.startsWith("/static")) {
+                contentLanguage = Locale.LanguageRange.parse(call.request.acceptLanguage())
+            }
+        }
+        get {
+            // landing page
+            call.respond(
+                ThymeleafContent(
+                    template = "landing",
+                    model = mapOf(
+                        "glRoute" to "/$GL_ROUTE"
+                    ),
+                    locale = getPreferredLocale(contentLanguage,"landing")
+                )
+            )
+        }
+    }
+}
